refactor: Nest user-feedback tools.
