+++ src/tools/index.ts
@@ -32,8 +32,8 @@ import "./recommendations";
 import "./report-forecasts";
 import "./resource-reports";
 import "./scenario-models";
-import "./submit-user-feedback";
 import "./tags";
 import "./update-report-notification";
+import "./user-feedback";
 import "./users";
 import "./workspaces";
