+++ src/tools/user-feedback/index.ts
@@ -1,0 +1,1 @@
+import "./submit-user-feedback";
